ajout badge codewars + video wild + fixes css + corrections coquilles + création d'un fichier projets
