+++ src/components/About.module.css
@@ -4,14 +4,27 @@
   align-items: flex-start;
 }
 
+.picBlock {
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+}
+
 .profilePic {
   border-radius: 100%;
   height: 20vw;
+  object-fit: contain;
   margin-right: 15px;
 }
 
 .profilePic:hover {
   animation: pic 10s infinite linear;
+}
+
+.badge {
+  width: 70%;
+  max-width: 500px;
+  margin: auto;
 }
 
 @keyframes pic {
